requirements.txt
colorama
absl-py==0.9.0
astor==0.8.1
bert-tensorflow==1.0.1
Click==7.0
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.18.1
opt-einsum==3.1.0
prettytable==0.7.2
protobuf==3.11.3
regex==2020.2.20
sacremoses==0.0.38
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
tqdm==4.43.0
Werkzeug==1.0.0
wrapt==1.12.0
